refactor(channels): consolidate flat streaming fallback into the deprecation module
