+++ src/channels/streaming.ts
@@ -5,18 +5,19 @@ import { normalizeTrimmedStringList } from "@openclaw/normalization-core/string-
 import { formatToolDetail, resolveToolDisplay } from "../agents/tool-display.js";
 import { formatToolAggregate } from "../auto-reply/tool-meta.js";
 import type {
-  BlockStreamingChunkConfig,
-  BlockStreamingCoalesceConfig,
   ChannelStreamingCommandTextMode,
   ChannelStreamingProgressConfig,
   ChannelStreamingConfig,
   StreamingMode,
-  TextChunkMode,
 } from "../config/types.base.js";
 import { asBoolean } from "../utils/boolean.js";
-import {
-  warnFlatStreamingKeyFallback,
-  warnScalarStreamingFallback,
+import { warnScalarStreamingFallback } from "./streaming-flat-key-deprecation.js";
+
+export {
+  resolveChannelStreamingChunkMode,
+  resolveChannelStreamingBlockEnabled,
+  resolveChannelStreamingBlockCoalesce,
+  resolveChannelStreamingPreviewChunk,
 } from "./streaming-flat-key-deprecation.js";
 
 export type {
@@ -44,24 +45,15 @@ export type StreamingCompatEntry = {
   draftChunk?: unknown;
 };
 
-// Nested streaming config wins. Every bundled channel now uses a nested-only
-// streaming schema with doctor migrating the flat spellings, so in-tree the
-// flat delivery keys (chunkMode, blockStreaming, blockStreamingCoalesce,
-// draftChunk) are legacy config. Those reads and the scalar `streaming` read in
-// resolveChannelPreviewStreamMode serve external SDK plugin configs only and
-// emit once-per-key deprecation warnings. Remove them (and the flat
-// StreamingCompatEntry fields) when the next release train closes the SDK
-// deprecation window.
+// Nested config wins; flat delivery fallback lives wholly in streaming-flat-key-deprecation.ts.
+// After the next release train, delete that module and re-exports, scalar read in
+// resolveChannelPreviewStreamMode, and flat StreamingCompatEntry fields.
 // Mode-family aliases (streamMode) are doctor-only and stay unread here.
 
 function asObjectRecord(value: unknown): Record<string, unknown> | null {
   return value && typeof value === "object" && !Array.isArray(value)
     ? (value as Record<string, unknown>)
     : null;
-}
-
-function asTextChunkMode(value: unknown): TextChunkMode | undefined {
-  return value === "length" || value === "newline" ? value : undefined;
 }
 
 function asInteger(value: unknown): number | undefined {
@@ -87,14 +79,6 @@ function parsePreviewStreamingMode(value: unknown): StreamingMode | null {
     return normalized;
   }
   return null;
-}
-
-function asBlockStreamingCoalesceConfig(value: unknown): BlockStreamingCoalesceConfig | undefined {
-  return (asObjectRecord(value) as BlockStreamingCoalesceConfig | null) ?? undefined;
-}
-
-function asBlockStreamingChunkConfig(value: unknown): BlockStreamingChunkConfig | undefined {
-  return (asObjectRecord(value) as BlockStreamingChunkConfig | null) ?? undefined;
 }
 
 function asProgressConfig(value: unknown): ChannelStreamingProgressConfig | undefined {
@@ -779,65 +763,6 @@ export function getChannelStreamingConfigObject(
 ): ChannelStreamingConfig | undefined {
   const streaming = asObjectRecord(entry?.streaming);
   return streaming ? (streaming as ChannelStreamingConfig) : undefined;
-}
-
-function resolveWithFlatFallback<T>(params: {
-  nested: T | undefined;
-  flat: T | undefined;
-  flatKey: string;
-  nestedPath: string;
-}): T | undefined {
-  if (params.nested !== undefined) {
-    return params.nested;
-  }
-  if (params.flat !== undefined) {
-    warnFlatStreamingKeyFallback(params.flatKey, params.nestedPath);
-  }
-  return params.flat;
-}
-
-export function resolveChannelStreamingChunkMode(
-  entry: StreamingCompatEntry | null | undefined,
-): TextChunkMode | undefined {
-  return resolveWithFlatFallback({
-    nested: asTextChunkMode(getChannelStreamingConfigObject(entry)?.chunkMode),
-    flat: asTextChunkMode(entry?.chunkMode),
-    flatKey: "chunkMode",
-    nestedPath: "chunkMode",
-  });
-}
-
-export function resolveChannelStreamingBlockEnabled(
-  entry: StreamingCompatEntry | null | undefined,
-): boolean | undefined {
-  return resolveWithFlatFallback({
-    nested: asBoolean(getChannelStreamingConfigObject(entry)?.block?.enabled),
-    flat: asBoolean(entry?.blockStreaming),
-    flatKey: "blockStreaming",
-    nestedPath: "block.enabled",
-  });
-}
-
-export function resolveChannelStreamingBlockCoalesce(
-  entry: StreamingCompatEntry | null | undefined,
-): BlockStreamingCoalesceConfig | undefined {
-  return resolveWithFlatFallback({
-    nested: asBlockStreamingCoalesceConfig(getChannelStreamingConfigObject(entry)?.block?.coalesce),
-    flat: asBlockStreamingCoalesceConfig(entry?.blockStreamingCoalesce),
-    flatKey: "blockStreamingCoalesce",
-    nestedPath: "block.coalesce",
-  });
-}
-
-export function resolveChannelStreamingPreviewChunk(
-  entry: StreamingCompatEntry | null | undefined,
-): BlockStreamingChunkConfig | undefined {
-  return resolveWithFlatFallback({
-    nested: asBlockStreamingChunkConfig(getChannelStreamingConfigObject(entry)?.preview?.chunk),
-    flat: asBlockStreamingChunkConfig(entry?.draftChunk),
-    flatKey: "draftChunk",
-    nestedPath: "preview.chunk",
-  });
 }
 
 export function resolveChannelStreamingPreviewToolProgress(
